refactor: Footer 컴포넌트 내부 분리
