+++ src/components/footer/index.jsx
@@ -1,30 +1,8 @@
-import { Link } from "react-router";
-
-// Imges
-import facebook from "../../assets/icons/ic_facebook.svg";
-import twitter from "../../assets/icons/ic_twitter.svg";
-import youtube from "../../assets/icons/ic_youtube.svg";
-import instagram from "../../assets/icons/ic_instagram.svg";
-
 // Components
 import FooterLayout from "../layout/FooterLayout";
-
-const SNS_DATA = [
-  {
-    id: "facebook",
-    href: "https://facebook.com",
-    src: facebook,
-    alt: "페이스북",
-  },
-  { id: "twitter", href: "https://twitter.com", src: twitter, alt: "트위터" },
-  { id: "youtube", href: "https://youtube.com", src: youtube, alt: "유튜브" },
-  {
-    id: "instagram",
-    href: "https://instagram.com",
-    src: instagram,
-    alt: "인스타그램",
-  },
-];
+import FooterCopyright from "./FooterCopyright";
+import FooterLinks from "./FooterLinks";
+import FooterSNS from "./FooterSNS";
 
 export default function Footer() {
   return (
@@ -49,44 +27,3 @@ const FooterContainer = ({ children }) => {
     </div>
   );
 };
-
-const FooterCopyright = ({ children, className }) => (
-  <span
-    className={`text-gray-400 font-pretendard text-base leading-4.75 ${className}`}
-  >
-    {children}
-  </span>
-);
-
-const FooterLinks = ({ className }) => (
-  <div
-    className={`text-gray-200 font-pretendard text-base leading-4.75 ${className}`}
-  >
-    <Link to="/privacy" className="hover:underline">
-      Privacy Policy
-    </Link>
-    <Link to="/faq" className="hover:underline">
-      FAQ
-    </Link>
-  </div>
-);
-
-const FooterSNS = ({ className }) => (
-  <ul className={className}>
-    {SNS_DATA.map((sns) => (
-      <SNSItem key={sns.id} {...sns} />
-    ))}
-  </ul>
-);
-
-const SNSItem = ({ href, src, alt }) => (
-  <li>
-    <a href={href} target="_blank" rel="noopener noreferrer">
-      <img
-        src={src}
-        alt={`${alt} 바로가기`}
-        className="w-5 h-5 lg:w-6 lg:h-6"
-      />
-    </a>
-  </li>
-);
